+++ src/main/java/com/quasiris/qsf/query/parser/QsfqlParser.java
@@ -172,6 +172,11 @@ public class QsfqlParser {
                     searchFilter.setFilterType(FilterType.TERM);
                     searchFilter.setFilterOperator(FilterOperator.OR);
                     query.getSearchFilterList().add(searchFilter);
+                } else if (".not".equals(filterType)) {
+                        SearchFilter searchFilter = createSearchFilter(filterName, filterValues);
+                        searchFilter.setFilterType(FilterType.TERM);
+                        searchFilter.setFilterOperator(FilterOperator.NOT);
+                        query.getSearchFilterList().add(searchFilter);
                 } else if (".slider".equals(filterType)) {
                     SearchFilter searchFilter = createRangeFilter(filterName, filterValues);
                     searchFilter.setFilterType(FilterType.SLIDER);
@@ -185,7 +190,7 @@ public class QsfqlParser {
                 } else if (".tree".equals(filterType)) {
                     List<SearchFilter> searchFilters = createTreeFilter(filterName, filterValues);
                     query.getSearchFilterList().addAll(searchFilters);
-            }
+                }
 
             }
         }